Reworked the MyRatingsActivity.java TextView where the rated books are displayed via an inner private class that is essentially a re-skinned ArrayList.  Added an app icon, so the app can be displayed in the apps folder with the new icon (and not the default).  Added non-null notifications to all applicable methods in the RatingsDBHelper.java file, as well as tested for empty strings passed as a parameter.  Fixed a few silent data type errors in RatingActivity.java, specifically regarding the RatingBar object, and fixed a few silent data type mismatches for the ISBN (which started as an integer, but was too big, so I had to change it to a String).

+++ app/src/main/java/it372/finalprojmcgrath/RatingActivity.java
@@ -118,8 +118,8 @@ public class RatingActivity extends AppCompatActivity {
                 spnGenre.setSelection(spinnerPosition);
             }
             // restore rating
-            int tempRating = savedInstanceState.getInt("rating");
-            ratingBar.setRating(tempRating);
+            double tempRating = savedInstanceState.getDouble("rating");
+            ratingBar.setRating((float) tempRating);
             // restore user's thoughts
             edtUsersThoughts.setText(savedInstanceState.getString("thoughts"));
         }
